Authentication: Login & Registration page
This commit implements the registration and login pages. The registration page includes a name field, an email field, and a password field. The login page only has the email and the password input fields. This commit also connects to the firebase and uses firebase authentication to implement email/password based validation.