Home Page: Navbar, Banner/Slider, Event Category, Footer
This commit implements 4 Home Page sections. The navbar so far only includes the home, login, and the register navigations. The banner section is a slider. The Event Category section displays 6 event cards as of now, and the Footer section is a classic footer. This commit applies AOS(Animation On Scroll) on the cards in the event category section and on the footer. The event category cards have a button that takes the user to a private page about the event's details and descriptions.