Service Description Page: Video, private route
This commit implements the private route for 'Service Description' Page. The page has the navbar at the top and the footer at the bottom. these two sections are shared between home and this page so far. After the navbar, there's a video related to the specific event service, after that there's an event tile followed by a picture on the left and a short description on the right. This commit also makes this route private, so only logged in users are able to navigate to this page. If a user is not logged in they are redirected to the login page.